# Various Ruby driver optimizations