--- a/lib/redis_client/ruby_connection/buffered_io.rb
+++ b/lib/redis_client/ruby_connection/buffered_io.rb
@@ -12,7 +12,7 @@ class RedisClient
 
       def initialize(io, read_timeout:, write_timeout:, chunk_size: 4096)
         @io = io
-        @buffer = "".b
+        @buffer = "".dup.force_encoding(Encoding.default_external)
         @offset = 0
         @chunk_size = chunk_size
         @read_timeout = read_timeout