I recently paired with @tenderlove on trying to see if the Ruby vs Hiredis performance difference could be closed.

We came up with a bunch of small changes that really close the gap quite significantly, especially when YJIT is enabled.

The Ruby driver is now either on par with hiredis, or just marginally slower on some cases.

There are a few more optimization we thought of, but they require some changes in Ruby, I'll try to work on these this year. Notably a way to create hashes with a given capacity would help a lot.

## Before (YJIT)

ruby: `ruby 3.3.0 (2023-12-25 revision 5124f9ac75) [arm64-darwin23]`

redis-server: `Redis server v=7.0.12 sha=00000000:0 malloc=libc bits=64 build=a11d0151eabf466c`


### small string x 100

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:     6745.7 i/s
                ruby:     5182.0 i/s - 1.30x  slower

```

### large string x 100

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:      303.5 i/s
                ruby:      343.1 i/s - 1.13x  faster

```

### small list x 100

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:     3683.6 i/s
                ruby:     1952.3 i/s - 1.89x  slower

```

### large list

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:     6540.2 i/s
                ruby:     2710.0 i/s - 2.41x  slower

```

### small hash x 100

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:     4002.6 i/s
                ruby:     2317.0 i/s - 1.73x  slower

```

### large hash

```
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
             hiredis:     2467.1 i/s
                ruby:     2439.0 i/s - same-ish: difference falls within error

```

## After (YJIT)

ruby: `ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) [arm64-darwin23]`

redis-server: `Redis server v=7.0.12 sha=00000000:0 malloc=libc bits=64 build=a11d0151eabf466c`


### small string x 100

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:     6407.8 i/s
                ruby:     5852.0 i/s - same-ish: difference falls within error

```

### large string x 100

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:      302.8 i/s
                ruby:      337.3 i/s - same-ish: difference falls within error

```

### small list x 100

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:     4067.7 i/s
                ruby:     2721.5 i/s - 1.49x  slower

```

### large list

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:     7138.7 i/s
                ruby:     6605.4 i/s - same-ish: difference falls within error

```

### small hash x 100

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:     4219.8 i/s
                ruby:     3586.4 i/s - 1.18x  slower

```

### large hash

```
ruby 3.4.0dev (2024-03-19T14:18:56Z master 5c2937733c) +YJIT [arm64-darwin23]
             hiredis:     5240.9 i/s
                ruby:     5312.5 i/s - same-ish: difference falls within error

```

NB: the later use `ruby-head` to benefit from some YJIT opts.