@casperisfine 2024-03-22
The TruffleRuby failure seem legit and may be indicative of a TruffleRuby bug:

```
RedisClientTest#test_encoding [test/redis_client_test.rb:31]:
Expected: #<Encoding:UTF-8>
  Actual: #<Encoding:ASCII-8BIT>
```

I'll dig this one down and try to come up with a minimal repro. FYI @eregon