@casperisfine 2024-03-22
Ruby spec for the TruffleRuby discrepancy: https://github.com/ruby/spec/pull/1145