@mperham 2024-03-22
21.1:

```
src/sidekiq % RUBY_YJIT_ENABLE=1 LATENCY=0 bin/sidekiqload
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
THREADS: nil, LATENCY: 0, AJ: nil, PROFILE: nil
2024-03-22T16:31:33.787Z pid=24663 tid=l27 ERROR: Setup RSS: 46496
2024-03-22T16:31:36.538Z pid=24663 tid=l27 WARN: Created 500000 jobs in 2.748508 sec
2024-03-22T16:31:36.538Z pid=24663 tid=l27 WARN: Starting load
2024-03-22T16:31:36.538Z pid=24663 tid=l27 ERROR: Simulating 0ms of latency between Sidekiq and redis
2024-03-22T16:31:53.557Z pid=24663 tid=lj3 ERROR: Done, 500000 jobs in **17.018055** sec, 29380 jobs/sec
2024-03-22T16:31:53.563Z pid=24663 tid=lj3 ERROR: Ending RSS: 54400
2024-03-22T16:31:53.564Z pid=24663 tid=lj3 ERROR: Now here's the latency for three jobs
0.0007669925689697266
0.00015878677368164062
0.00014734268188476562
src/sidekiq % RUBY_YJIT_ENABLE=1 LATENCY=0 bin/sidekiqload
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
THREADS: nil, LATENCY: 0, AJ: nil, PROFILE: nil
2024-03-22T16:31:57.331Z pid=24700 tid=l2s ERROR: Setup RSS: 45648
2024-03-22T16:32:00.203Z pid=24700 tid=l2s WARN: Created 500000 jobs in 2.869176 sec
2024-03-22T16:32:00.203Z pid=24700 tid=l2s WARN: Starting load
2024-03-22T16:32:00.203Z pid=24700 tid=l2s ERROR: Simulating 0ms of latency between Sidekiq and redis
2024-03-22T16:32:17.020Z pid=24700 tid=lhw ERROR: Done, 500000 jobs in **16.816453** sec, 29732 jobs/sec
2024-03-22T16:32:17.027Z pid=24700 tid=lhw ERROR: Ending RSS: 53104
2024-03-22T16:32:17.027Z pid=24700 tid=lhw ERROR: Now here's the latency for three jobs
0.0009238719940185547
0.00021910667419433594
0.00011897087097167969
```

This branch:

```
src/sidekiq % RUBY_YJIT_ENABLE=1 LATENCY=0 bin/sidekiqload
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
THREADS: nil, LATENCY: 0, AJ: nil, PROFILE: nil
2024-03-22T16:29:32.030Z pid=24563 tid=gu3 ERROR: Setup RSS: 46432
2024-03-22T16:29:34.784Z pid=24563 tid=gu3 WARN: Created 500000 jobs in 2.736431 sec
2024-03-22T16:29:34.784Z pid=24563 tid=gu3 WARN: Starting load
2024-03-22T16:29:34.785Z pid=24563 tid=gu3 ERROR: Simulating 0ms of latency between Sidekiq and redis
2024-03-22T16:29:51.390Z pid=24563 tid=ggr ERROR: Done, 500000 jobs in **16.603711** sec, 30113 jobs/sec
2024-03-22T16:29:51.397Z pid=24563 tid=ggr ERROR: Ending RSS: 53600
2024-03-22T16:29:51.398Z pid=24563 tid=ggr ERROR: Now here's the latency for three jobs
0.0010797977447509766
0.0003361701965332031
0.00019621849060058594
src/sidekiq % RUBY_YJIT_ENABLE=1 LATENCY=0 bin/sidekiqload
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
THREADS: nil, LATENCY: 0, AJ: nil, PROFILE: nil
2024-03-22T16:29:56.089Z pid=24584 tid=l34 ERROR: Setup RSS: 44496
2024-03-22T16:29:58.893Z pid=24584 tid=l34 WARN: Created 500000 jobs in 2.801765 sec
2024-03-22T16:29:58.893Z pid=24584 tid=l34 WARN: Starting load
2024-03-22T16:29:58.893Z pid=24584 tid=l34 ERROR: Simulating 0ms of latency between Sidekiq and redis
2024-03-22T16:30:15.505Z pid=24584 tid=lgg ERROR: Done, 500000 jobs in **16.608818** sec, 30104 jobs/sec
2024-03-22T16:30:15.521Z pid=24584 tid=lgg ERROR: Ending RSS: 52064
2024-03-22T16:30:15.521Z pid=24584 tid=lgg ERROR: Now here's the latency for three jobs
0.0014886856079101562
0.0005791187286376953
0.0002732276916503906
```

A very slight improvement to Sidekiq's benchmark.